[Improve] Remove paused status of Activity
This gives a lot of benefits. It seems like we don't really need a
paused status. We can just have a binary of in-progress and stopped.
This allows for better cleanup of the component. So when the Activity
component is mounted, everything is started and then when it unmounts
everything finishes. Since we will be stacking screens on top of each
other this is possible.